% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{items}
\alias{items}
\title{Stardew Valley items dataset}
\format{
A data frame with...
}
\source{
Where the data comes from (if applicable).
}
\usage{
data(items)
}
\description{
This dataset includes information about various Stardew Valley items.
}
\keyword{datasets}
